fix: resolve failing tests with pandas>=2.2
Fixes failing tests that were due to a change in
[pandas>=2.2](https://pandas.pydata.org/docs/reference/api/pandas.core.groupby.DataFrameGroupBy.apply.html)
related to using `.apply()` on grouped-by objects. Fixes a bunch of
warnings related to dubious type mixing within columns. I also finally
fixes the import issue due to using the same name everywhere and adds a
unit test for verify that the import is not broken again.

It also deprecates 3.8 since it's almost EOL.

+++ tests/test_io.py
@@ -240,7 +240,9 @@ def treatments_dict_rand_index():
             "stratum": [0] * 40 + [1] * 30 + [2] * 30,
         }
     )
-    data = data.set_index(pd.Index(np.random.choice(300, 100, replace=False)))
+    data = data.set_index(
+        pd.Index(np.random.choice(300, 100, replace=False)).astype(np.int64)
+    )
     idx_col = "id"
 
     treatments = stochatreat(
